`Link` component renamed to `HyperLink`
closes #627
