+++ packages/components/ui/src/index.js
@@ -15,9 +15,9 @@ export {default as Banner} from './Banner.svelte';
 export {default as CenteredView} from './CenteredView.svelte';
 export {default as CopyToClipboard} from './CopyToClipboard.svelte';
 export {default as HighlightedText} from './HighlightedText.svelte';
+export {default as HyperLink} from './HyperLink.svelte';
 export {default as Input} from './Input.svelte';
 export {default as LayoutHMF} from './LayoutHMF.svelte';
-export {default as Link} from './Link.svelte';
 export {default as LinkButton} from './LinkButton.svelte';
 export {default as LoadingView} from './LoadingView.svelte';
 export {default as MessageView} from './MessageView.svelte';
